requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Cors==3.0.10
python-dateutil==2.8.2
